prevent delete files created during sync

+++ src/renderer/logic/sync.js
@@ -331,6 +331,7 @@ function CreateLocalFolders () {
 function CleanLocalFolders () {
   return new Promise(async (resolve, reject) => {
     const localPath = await database.Get('xPath')
+    const syncDate = database.Get('syncStartDate')
     // Get a list of all local folders
     tree.GetLocalFolderList(localPath).then((list) => {
       // Check what items are in dbFolders
@@ -341,8 +342,14 @@ function CleanLocalFolders () {
             next()
           } else {
             // Should DELETE that folder in local
-            console.log('Delete folder', item)
-            rimraf(item, (err) => next(err))
+            const creationDate = fs.statSync(item)
+            // Delete only if
+            if (creationDate <= syncDate) {
+              console.log('Delete folder', item)
+              rimraf(item, (err) => next(err))
+            } else {
+              next()
+            }
           }
         }).catch(err => {
           console.log('ITEM ERR', err)
@@ -358,12 +365,16 @@ function CleanLocalFolders () {
 function CleanLocalFiles () {
   return new Promise(async (resolve, reject) => {
     const localPath = await database.Get('xPath')
+    const syncDate = database.Get('syncStartDate')
     tree.GetLocalFileList(localPath).then(list => {
       async.eachSeries(list, (item, next) => {
         database.FileGet(item).then(fileObj => {
           if (!fileObj) {
-            console.log('Delete file %s', item)
-            fs.unlinkSync(item)
+            const creationDate = fs.statSync(item)
+            if (creationDate <= syncDate) {
+              console.log('Delete file %s', item)
+              fs.unlinkSync(item)
+            }
             next()
           } else {
             next()
